+++ src/assets/styles/components.css
@@ -2,6 +2,6 @@
 @import "./components/login-register.css";
 @import "./components/logout.css";
 @import "./components/navs.css";
-@import "./components/dashboard.css";
+@import "./components/home.css";
 @import "./components/channels-chats.css";
 @import "./components/not-found.css";
